video_analysis_system/requirements.txt
flask>=2.3.0
requests>=2.31.0
openai>=1.3.0
pymongo>=4.6.0
moviepy>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
dnspython>=2.4.0
yt-dlp>=2023.11.16 